Shrink code download, add native ppc64le support
Make ubuntu-build-all.sh runnable on ppc64le.

I could not get "gn" working on ppc64le, so I worked around it by
replacing depot_tools operations with Git and GNU make.

With this change the WebRTC checkout is ~4.6 GiB instead of ~17 GiB due
to shallow clones.

The Makefile is used to build libdcsctp.a to be consumed by
libdcsctp4j.so.

Cross-compilers are used unconditionally so that, for example, when
ubuntu-build-all.sh is run on ppc64le, it cross-builds
linux-x86-64/libdcsctp4j.so and linux-aarch64/libdcsctp4j.so.  I do not
have Jitsi setups to test these.  However the natively-compiled
linux-ppc64le/libdcsctp4j.so does seem to work on my ppc64le Jitsi
setup.

I left the depot_tools arguments present but ignored in the scripts so
that the calling interface is unchanged, to allow for experimentation.